fix transacciones
